Add logs bloom
